requirements.txt
Django==4.2.7
Pillow==10.0.0
tensorflow==2.16.1
numpy==1.26.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7
